fix(slc): Renamed modules to *_
