+++ duui-slc/spacy/src/main/python/app/parser/spacy_.py
